requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
img2pdf==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langsmith==0.1.139
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
ocrmypdf==16.6.0
openai==1.53.0
orjson==3.10.11
packaging==24.1
pdfminer.six==20240706
pi_heif==0.20.0
pikepdf==9.4.0
pillow==11.0.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
pip==23.2.1
pluggy==1.5.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.6.1
Pygments==2.18.0
PyMuPDF==1.24.13
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==68.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.6
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
Werkzeug==3.1.1
wheel==0.41.2
wrapt==1.16.0
yarl==1.17.1
